# openapi.yaml
openapi: 3.0.3
info:
  title: Finance MCP Server
  description: |
    Real-Time Financial Data MCP Integration System.

    This server implements the Model Context Protocol (MCP) for financial data retrieval
    with support for multiple free data providers, caching layers, and graph lineage tracking.

    ## Features
    - **MCP Protocol Compliance**: Standard MCP endpoints for tool discovery and execution
    - **Multi-Source Data**: Alpha Vantage, Finnhub, and Binance WebSocket connectors
    - **Smart Caching**: Redis hot cache + Qdrant semantic cache
    - **Lineage Tracking**: Neo4j graph database for data lineage

    ## Free Data Sources
    - Alpha Vantage (5 calls/min)
    - Finnhub (60 calls/min)  
    - Binance WebSocket (unlimited crypto streaming)
  version: 1.0.0
  contact:
    name: Finance MCP Support
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Local development server

tags:
  - name: MCP Protocol
    description: Model Context Protocol endpoints
  - name: Subscriptions
    description: Stream subscription management
  - name: Health
    description: Server health and status

paths:
  /.well-known/mcp:
    get:
      tags:
        - MCP Protocol
      summary: MCP Server Metadata
      description: Returns server metadata and protocol information
      operationId: getMCPMetadata
      responses:
        "200":
          description: Server metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPMetadata"
              example:
                name: finance-mcp
                version: "1.0.0"
                protocol_version: "1.0"
                description: Real-Time Financial Data MCP Integration System
                capabilities:
                  - tools
                  - subscriptions
                endpoints:
                  capabilities: /capabilities
                  invoke: /invoke
                  subscribe: /subscribe
                  unsubscribe: /unsubscribe

  /capabilities:
    get:
      tags:
        - MCP Protocol
      summary: Get Available Tools
      description: Returns list of available MCP tools and connectors
      operationId: getCapabilities
      responses:
        "200":
          description: Capabilities list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"

  /invoke:
    post:
      tags:
        - MCP Protocol
      summary: Invoke MCP Tool
      description: |
        Execute an MCP tool with the provided arguments.

        ## Available Tools

        ### quote.latest
        Get the latest price quote for a financial instrument.

        ### quote.stream
        Subscribe to real-time price stream (crypto only via Binance).

        ### multi_agent_analysis
        Run orchestrated bull vs bear market reasoning with a judge synthesis.
      operationId: invokeTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolInvocation"
            examples:
              quoteLatest:
                summary: Get latest AAPL quote
                value:
                  tool_name: quote.latest
                  arguments:
                    symbol: AAPL
                    maxAgeSec: 60
                  agent_id: agent_001
                  query_text: What is the current price of Apple stock?
              quoteLatestCrypto:
                summary: Get latest Bitcoin price
                value:
                  tool_name: quote.latest
                  arguments:
                    symbol: BTCUSDT
                    maxAgeSec: 30
              quoteStream:
                summary: Subscribe to crypto stream
                value:
                  tool_name: quote.stream
                  arguments:
                    symbol: ETHUSDT
                    channel: trades
                  agent_id: agent_001
      responses:
        "200":
          description: Tool executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
              examples:
                success:
                  summary: Successful quote response
                  value:
                    success: true
                    data:
                      symbol: AAPL
                      price: 150.50
                      timestamp: "2025-11-29T12:00:00Z"
                      data_source: finnhub
                      cache_hit: false
                      latency_ms: 245.5
                      volume: 1500000
                      high: 152.00
                      low: 149.00
                    cache_hit: false
                    data_source: finnhub
                    latency_ms: 245.5
        "400":
          description: Invalid request or tool execution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
              example:
                success: false
                error: "Invalid symbol format: INVALID!!!"
        "500":
          description: Internal server error

  /subscribe:
    post:
      tags:
        - Subscriptions
      summary: Subscribe to Stream
      description: Subscribe to a real-time data stream for a symbol
      operationId: subscribe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequest"
            example:
              symbol: BTCUSDT
              channel: trades
              agent_id: agent_001
      responses:
        "200":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
              example:
                success: true
                data:
                  subscription_id: sub_abc12345
                  status: subscribed
                  symbol: BTCUSDT
                  channel: trades
        "400":
          description: Subscription failed

  /unsubscribe:
    post:
      tags:
        - Subscriptions
      summary: Unsubscribe from Stream
      description: Cancel an active stream subscription
      operationId: unsubscribe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeRequest"
            example:
              subscription_id: sub_abc12345
      responses:
        "200":
          description: Unsubscribed successfully
          content:
            application/json:
              example:
                success: true
                data:
                  subscription_id: sub_abc12345
                  status: unsubscribed
        "400":
          description: Unsubscribe failed

  /subscriptions:
    get:
      tags:
        - Subscriptions
      summary: List Active Subscriptions
      description: Get all currently active stream subscriptions
      operationId: listSubscriptions
      responses:
        "200":
          description: Active subscriptions list
          content:
            application/json:
              example:
                subscriptions:
                  sub_abc12345:
                    symbol: BTCUSDT
                    channel: trades
                    agent_id: agent_001

  /health:
    get:
      tags:
        - Health
      summary: Health Check
      description: Check server health and connectivity status
      operationId: healthCheck
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
              example:
                status: healthy
                redis_connected: true
                active_subscriptions: 2

components:
  schemas:
    MCPMetadata:
      type: object
      properties:
        name:
          type: string
          description: Server name
        version:
          type: string
          description: Server version
        protocol_version:
          type: string
          description: MCP protocol version
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        endpoints:
          type: object
          additionalProperties:
            type: string

    Capabilities:
      type: object
      properties:
        server:
          type: object
          properties:
            name:
              type: string
            version:
              type: string
            description:
              type: string
        tools:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        connectors:
          type: array
          items:
            $ref: "#/components/schemas/Connector"

    Tool:
      type: object
      properties:
        name:
          type: string
          description: Tool identifier
        description:
          type: string
        inputSchema:
          type: object
        outputSchema:
          type: object

    Connector:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [rest, websocket]
        description:
          type: string
        rate_limit:
          type: string

    ToolInvocation:
      type: object
      required:
        - tool_name
        - arguments
      properties:
        tool_name:
          type: string
          description: Name of the tool to invoke
          enum:
            - quote.latest
            - quote.stream
            - trace_impact
            - analyze_news_impact
            - multi_agent_analysis
        arguments:
          type: object
          description: Tool-specific arguments
        agent_id:
          type: string
          description: Optional agent identifier for lineage tracking
        query_text:
          type: string
          description: Original query text for semantic caching

    ToolResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        cache_hit:
          type: boolean
        data_source:
          type: string
          nullable: true
        latency_ms:
          type: number

    QuoteData:
      type: object
      properties:
        symbol:
          type: string
        price:
          type: number
        timestamp:
          type: string
          format: date-time
        data_source:
          type: string
          enum:
            - alpha_vantage
            - finnhub
            - binance
            - redis_cache
            - semantic_cache
        cache_hit:
          type: boolean
        latency_ms:
          type: number
        volume:
          type: number
          nullable: true
        high:
          type: number
          nullable: true
        low:
          type: number
          nullable: true
        open:
          type: number
          nullable: true
        previous_close:
          type: number
          nullable: true

    SubscriptionRequest:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
          description: Ticker symbol to subscribe to
        channel:
          type: string
          enum:
            - trades
            - quotes
          default: trades
        agent_id:
          type: string
          nullable: true

    SubscriptionResponse:
      type: object
      properties:
        subscription_id:
          type: string
        status:
          type: string
        symbol:
          type: string
        channel:
          type: string

    UnsubscribeRequest:
      type: object
      required:
        - subscription_id
      properties:
        subscription_id:
          type: string

    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum:
            - healthy
            - degraded
            - unhealthy
        redis_connected:
          type: boolean
        active_subscriptions:
          type: integer